+++ bootstrap.tf
@@ -3,6 +3,7 @@ module "bootstrap" {
   source              = "./modules/bootstrap"
   subnet_id           = var.subnets_ids[0]
   cluster_name        = aws_eks_cluster.this.name
+  ami                 = var.bootstrap_ami
   aws_account         = var.aws_account
   admin_role_arns     = var.admin_role_arns
   dev_role_arns       = var.dev_role_arns